fix: If field `_id` present in OutputType then make it NonNull
This fix makes work with Flowtype and Relay Modern more comfortable. It
makes _id field NonNull, so no need an additional checks on the client
side is _id null or not.

Mongoose and mongo almost always have _id for the document and it is not
null. In rare cases if you tune mongoose model to have _id optional, you
should manually change type for this generated field.